Made the Proposed Core API, with lib supporting it!
