+++ lib/memory/memory.js
@@ -314,7 +314,7 @@ class WasmBoyMemoryService {
   }
 
   // Function to reset stateful sections of memory
-  resetMemory() {
+  resetState() {
     for (let i = 0; i <= WASMBOY_GAME_BYTES_LOCATION; i++) {
       this.wasmByteMemory[i] = 0;
     }
@@ -329,9 +329,6 @@ class WasmBoyMemoryService {
         this.wasmByteMemory[WASMBOY_GAME_BYTES_LOCATION + i] = gameBytes[i];
       }
     }
-
-    // TODO: Handle getting a boot rom
-    this.wasmInstance.exports.initialize(isGbcEnabled ? 1 : 0, 0);
 
     this.loadedCartridgeMemoryState.ROM = true;
   }
@@ -508,14 +505,6 @@ class WasmBoyMemoryService {
           });
       }
     });
-  }
-
-  // Function to reset the state of the wasm core
-  resetState(isGbcEnabled, bootRom) {
-    this.resetMemory();
-
-    // TODO: Handle getting a boot rom
-    this.wasmInstance.exports.initialize(isGbcEnabled ? 1 : 0, 0);
   }
 
   // Function to return the current cartridge object
