+++ lib/wasmboy.js
@@ -1,18 +1,18 @@
-import fetch from "unfetch";
-import Promise from "promise-polyfill";
-import wasmModuleUrl from "../dist/core/index.untouched.wasm";
-import { WasmBoyGraphics } from "./graphics/graphics";
-import { WasmBoyAudio } from "./audio/audio";
-import { WasmBoyController } from "./controller/controller";
-import { WasmBoyMemory } from "./memory/memory";
+import fetch from 'unfetch';
+import Promise from 'promise-polyfill';
+import wasmModuleUrl from '../dist/core/index.untouched.wasm';
+import { WasmBoyGraphics } from './graphics/graphics';
+import { WasmBoyAudio } from './audio/audio';
+import { WasmBoyController } from './controller/controller';
+import { WasmBoyMemory } from './memory/memory';
 
 // requestAnimationFrame() for headless mode
-const raf = require("raf");
+const raf = require('raf');
 
 // Function to get performance timestamp
 // This is to support node vs. Browser
 const getPerformanceTimestamp = () => {
-  if (typeof window !== "undefined") {
+  if (typeof window !== 'undefined') {
     return performance.now();
   }
   return Date.now();
@@ -108,15 +108,9 @@ class WasmBoyLib {
       // Pause the game in case it was running
       this.pauseGame().then(() => {
         // Get our promises
-        const initPromises = [
-          this._fetchGameAsByteArray(game),
-          this._getWasmInstance()
-        ];
+        const initPromises = [this._fetchGameAsByteArray(game), this._getWasmInstance()];
 
-        if (
-          !this.headless &&
-          WasmBoyMemory.getLoadedCartridgeMemoryState().RAM
-        ) {
+        if (!this.headless && WasmBoyMemory.getLoadedCartridgeMemoryState().RAM) {
           initPromises.push(WasmBoyMemory.saveCartridgeRam());
         }
 
@@ -127,18 +121,28 @@ class WasmBoyLib {
 
             // Check if we are running headless
             if (this.headless) {
-              WasmBoyMemory.initializeHeadless(
-                this.wasmInstance,
-                this.wasmByteMemory,
-                this.saveStateCallback
-              );
+              WasmBoyMemory.initializeHeadless(this.wasmInstance, this.wasmByteMemory, this.saveStateCallback);
 
               // Clear what is currently in memory, then load the cartridge memory
               WasmBoyMemory.clearMemory();
               WasmBoyMemory.resetState();
 
               // TODO: Handle passing a boot rom
-              WasmBoyMemory.loadCartridgeRom(responses[0], false, false);
+              WasmBoyMemory.loadCartridgeRom(responses[0]);
+
+              // Run our initialization on the core
+              this.wasmInstance.exports.config(
+                0, // TODO: Include Boot Rom
+                this.isGbcEnabled ? 1 : 0,
+                this.audioBatchProcessing ? 1 : 0,
+                this.graphicsBatchProcessing ? 1 : 0,
+                this.timersBatchProcessing ? 1 : 0,
+                this.graphicsDisableScanlineRendering ? 1 : 0,
+                this.audioAccumulateSamples ? 1 : 0,
+                this.tileRendering ? 1 : 0,
+                this.tileCaching ? 1 : 0
+              );
+
               this.ready = true;
 
               resolve();
@@ -146,18 +150,10 @@ class WasmBoyLib {
               // Finally intialize all of our services
               // Initialize our services
               Promise.all([
-                WasmBoyGraphics.initialize(
-                  this.canvasElement,
-                  this.wasmInstance,
-                  this.wasmByteMemory
-                ),
+                WasmBoyGraphics.initialize(this.canvasElement, this.wasmInstance, this.wasmByteMemory),
                 WasmBoyAudio.initialize(this.wasmInstance, this.wasmByteMemory),
                 WasmBoyController.initialize(this.wasmInstance),
-                WasmBoyMemory.initialize(
-                  this.wasmInstance,
-                  this.wasmByteMemory,
-                  this.saveStateCallback
-                )
+                WasmBoyMemory.initialize(this.wasmInstance, this.wasmByteMemory, this.saveStateCallback)
               ])
                 .then(() => {
                   // Clear what is currently in memory, then load the carttridge memory
@@ -165,10 +161,19 @@ class WasmBoyLib {
                   WasmBoyMemory.resetState();
 
                   // TODO: Handle passing a boot rom
-                  WasmBoyMemory.loadCartridgeRom(
-                    responses[0],
-                    this.isGbcEnabled,
-                    false
+                  WasmBoyMemory.loadCartridgeRom(responses[0]);
+
+                  // Run our initialization on the core
+                  this.wasmInstance.exports.config(
+                    0, // TODO: Include Boot Rom
+                    this.isGbcEnabled ? 1 : 0,
+                    this.audioBatchProcessing ? 1 : 0,
+                    this.graphicsBatchProcessing ? 1 : 0,
+                    this.timersBatchProcessing ? 1 : 0,
+                    this.graphicsDisableScanlineRendering ? 1 : 0,
+                    this.audioAccumulateSamples ? 1 : 0,
+                    this.tileRendering ? 1 : 0,
+                    this.tileCaching ? 1 : 0
                   );
 
                   // Load the game's cartridge ram
@@ -197,19 +202,6 @@ class WasmBoyLib {
   reset(wasmBoyOptions) {
     this.initialize(this.canvasElement, wasmBoyOptions);
     WasmBoyMemory.resetState();
-    if (this.wasmInstance) {
-      // Run our initialization on the core
-      this.wasmInstance.exports.config(
-        this.audioBatchProcessing ? 1 : 0,
-        this.graphicsBatchProcessing ? 1 : 0,
-        this.timersBatchProcessing ? 1 : 0,
-        this.graphicsDisableScanlineRendering ? 1 : 0,
-        this.audioAccumulateSamples ? 1 : 0,
-        this.tileRendering ? 1 : 0,
-        this.tileCaching ? 1 : 0
-      );
-    }
-
     // Reload the game if one was already loaded
     if (this.loadedGame && !this.headless) {
       this.loadGame(this.loadedGame);
@@ -354,12 +346,11 @@ class WasmBoyLib {
       !this.headless &&
       !this.pauseFpsThrottle &&
       this.isAudioEnabled &&
-      this.wasmInstance.exports.getAudioQueueIndex() >
-        6000 * (this.gameboyFrameRate / 120) &&
+      this.wasmInstance.exports.getAudioQueueIndex() > 6000 * (this.gameboyFrameRate / 120) &&
       this.gameboyFrameRate <= 60
     ) {
       // TODO: Waiting for time stretching to resolve may be causing this
-      console.log("Waiting for audio...");
+      console.log('Waiting for audio...');
       return true;
     }
 
@@ -367,30 +358,19 @@ class WasmBoyLib {
     let response = this.wasmInstance.exports.update();
 
     // Handle our update() response
-    if (response > 0) {
+    if (response >= 0) {
       // See: wasm/cpu/opcodes update() function
-      // 1 = render a frame
-      // 2 = replace boot rom
-      // TODO: Find what should go here
+      // 0 = render a frame
       switch (response) {
-        case 1:
-        case 2:
+        case 0:
           break;
       }
 
       return true;
     } else {
-      console.log("Wasmboy Crashed!");
-      console.log(
-        `Program Counter: 0x${this.wasmInstance.exports
-          .getProgramCounter()
-          .toString(16)}`
-      );
-      console.log(
-        `Opcode: 0x${this.wasmByteMemory[
-          this.wasmInstance.exports.getProgramCounter()
-        ].toString(16)}`
-      );
+      console.log('Wasmboy Crashed!');
+      console.log(`Program Counter: 0x${this.wasmInstance.exports.getProgramCounter().toString(16)}`);
+      console.log(`Opcode: 0x${this.wasmByteMemory[this.wasmInstance.exports.getProgramCounter()].toString(16)}`);
       this.pauseGame();
       return false;
     }
@@ -440,104 +420,84 @@ class WasmBoyLib {
     return new Promise((resolve, reject) => {
       // Get our wasm instance from our wasmModule
       // https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/WebAssembly/instantiateStreaming
-      const memoryBase = WebAssembly.instantiateStreaming(
-        fetch(wasmModuleUrl),
-        {
-          env: {
-            log: (message, arg0, arg1, arg2, arg3, arg4, arg5) => {
-              // Grab our string
-              var len = new Uint32Array(
-                this.wasmInstance.exports.memory.buffer,
-                message,
-                1
-              )[0];
-              var str = String.fromCharCode.apply(
-                null,
-                new Uint16Array(
-                  this.wasmInstance.exports.memory.buffer,
-                  message + 4,
-                  len
-                )
-              );
-              if (arg0 !== -9999) str = str.replace("$0", arg0);
-              if (arg1 !== -9999) str = str.replace("$1", arg1);
-              if (arg2 !== -9999) str = str.replace("$2", arg2);
-              if (arg3 !== -9999) str = str.replace("$3", arg3);
-              if (arg4 !== -9999) str = str.replace("$4", arg4);
-              if (arg5 !== -9999) str = str.replace("$5", arg5);
+      const memoryBase = WebAssembly.instantiateStreaming(fetch(wasmModuleUrl), {
+        env: {
+          log: (message, arg0, arg1, arg2, arg3, arg4, arg5) => {
+            // Grab our string
+            var len = new Uint32Array(this.wasmInstance.exports.memory.buffer, message, 1)[0];
+            var str = String.fromCharCode.apply(null, new Uint16Array(this.wasmInstance.exports.memory.buffer, message + 4, len));
+            if (arg0 !== -9999) str = str.replace('$0', arg0);
+            if (arg1 !== -9999) str = str.replace('$1', arg1);
+            if (arg2 !== -9999) str = str.replace('$2', arg2);
+            if (arg3 !== -9999) str = str.replace('$3', arg3);
+            if (arg4 !== -9999) str = str.replace('$4', arg4);
+            if (arg5 !== -9999) str = str.replace('$5', arg5);
 
-              console.log("[WasmBoy] " + str);
-            },
-            hexLog: (arg0, arg1, arg2, arg3, arg4, arg5) => {
-              if (!this.logRequest) {
-                // Grab our arguments, and log as hex
-                let logString = "[WasmBoy]";
-                if (arg0 !== -9999) logString += ` 0x${arg0.toString(16)} `;
-                if (arg1 !== -9999) logString += ` 0x${arg1.toString(16)} `;
-                if (arg2 !== -9999) logString += ` 0x${arg2.toString(16)} `;
-                if (arg3 !== -9999) logString += ` 0x${arg3.toString(16)} `;
-                if (arg4 !== -9999) logString += ` 0x${arg4.toString(16)} `;
-                if (arg5 !== -9999) logString += ` 0x${arg5.toString(16)} `;
+            console.log('[WasmBoy] ' + str);
+          },
+          hexLog: (arg0, arg1, arg2, arg3, arg4, arg5) => {
+            if (!this.logRequest) {
+              // Grab our arguments, and log as hex
+              let logString = '[WasmBoy]';
+              if (arg0 !== -9999) logString += ` 0x${arg0.toString(16)} `;
+              if (arg1 !== -9999) logString += ` 0x${arg1.toString(16)} `;
+              if (arg2 !== -9999) logString += ` 0x${arg2.toString(16)} `;
+              if (arg3 !== -9999) logString += ` 0x${arg3.toString(16)} `;
+              if (arg4 !== -9999) logString += ` 0x${arg4.toString(16)} `;
+              if (arg5 !== -9999) logString += ` 0x${arg5.toString(16)} `;
 
-                // Uncomment to unthrottle
-                //console.log(logString);
+              // Uncomment to unthrottle
+              //console.log(logString);
 
-                // Comment the lines below to disable throttle
-                this.logRequest = true;
+              // Comment the lines below to disable throttle
+              this.logRequest = true;
+              setTimeout(() => {
+                console.log(logString);
+                this.logRequest = false;
+              }, Math.floor(Math.random() * 500));
+            }
+          },
+          performanceTimestamp: (id, value) => {
+            if (id === -9999) {
+              id = 0;
+            }
+
+            if (value === -9999) {
+              value = 0;
+            }
+
+            if (!this.performanceTimestamps[id]) {
+              this.performanceTimestamps[id] = {};
+              this.performanceTimestamps[id].throttle = false;
+              this.performanceTimestamps[id].totalTime = 0;
+              this.performanceTimestamps[id].value = 0;
+            }
+            if (!this.performanceTimestamps[id].throttle) {
+              if (this.performanceTimestamps[id].timestamp) {
+                // sleep a millisecond for hopefully more accurate times
+                let endTime = getPerformanceTimestamp();
+                let timeDifference = endTime - this.performanceTimestamps[id].timestamp;
+                this.performanceTimestamps[id].throttle = true;
+                this.performanceTimestamps[id].totalTime += timeDifference;
+                console.log(
+                  `[WasmBoy] Performance Timestamp. ID: ${id}, Time: ${timeDifference}, value difference: ${value -
+                    this.performanceTimestamps[id].value}, total time: ${this.performanceTimestamps[id].totalTime}`
+                );
+                this.performanceTimestamps[id].timestamp = false;
                 setTimeout(() => {
-                  console.log(logString);
-                  this.logRequest = false;
-                }, Math.floor(Math.random() * 500));
-              }
-            },
-            performanceTimestamp: (id, value) => {
-              if (id === -9999) {
-                id = 0;
-              }
-
-              if (value === -9999) {
-                value = 0;
-              }
-
-              if (!this.performanceTimestamps[id]) {
-                this.performanceTimestamps[id] = {};
-                this.performanceTimestamps[id].throttle = false;
-                this.performanceTimestamps[id].totalTime = 0;
-                this.performanceTimestamps[id].value = 0;
-              }
-              if (!this.performanceTimestamps[id].throttle) {
-                if (this.performanceTimestamps[id].timestamp) {
-                  // sleep a millisecond for hopefully more accurate times
-                  let endTime = getPerformanceTimestamp();
-                  let timeDifference =
-                    endTime - this.performanceTimestamps[id].timestamp;
-                  this.performanceTimestamps[id].throttle = true;
-                  this.performanceTimestamps[id].totalTime += timeDifference;
-                  console.log(
-                    `[WasmBoy] Performance Timestamp. ID: ${id}, Time: ${timeDifference}, value difference: ${value -
-                      this.performanceTimestamps[id].value}, total time: ${
-                      this.performanceTimestamps[id].totalTime
-                    }`
-                  );
-                  this.performanceTimestamps[id].timestamp = false;
-                  setTimeout(() => {
-                    this.performanceTimestamps[id].throttle = false;
-                  }, 100);
-                } else {
-                  this.performanceTimestamps[
-                    id
-                  ].timestamp = getPerformanceTimestamp();
-                  this.performanceTimestamps[id].value = value;
-                }
+                  this.performanceTimestamps[id].throttle = false;
+                }, 100);
+              } else {
+                this.performanceTimestamps[id].timestamp = getPerformanceTimestamp();
+                this.performanceTimestamps[id].value = value;
               }
             }
           }
         }
-      )
+      })
         .then(instantiatedWasm => {
           // Using || since rollup and webpack wasm loaders will return differently
-          const instance = (this.wasmInstance =
-            instantiatedWasm.instance || instantiatedWasm);
+          const instance = (this.wasmInstance = instantiatedWasm.instance || instantiatedWasm);
           const module = instantiatedWasm.module;
 
           // Get our memory from our wasm instance
@@ -552,20 +512,7 @@ class WasmBoyLib {
           // }
 
           // Will stay in sync
-          this.wasmByteMemory = new Uint8Array(
-            this.wasmInstance.exports.memory.buffer
-          );
-
-          // Run our initialization on the core
-          this.wasmInstance.exports.config(
-            this.audioBatchProcessing ? 1 : 0,
-            this.graphicsBatchProcessing ? 1 : 0,
-            this.timersBatchProcessing ? 1 : 0,
-            this.graphicsDisableScanlineRendering ? 1 : 0,
-            this.audioAccumulateSamples ? 1 : 0,
-            this.tileRendering ? 1 : 0,
-            this.tileCaching ? 1 : 0
-          );
+          this.wasmByteMemory = new Uint8Array(this.wasmInstance.exports.memory.buffer);
           resolve(this.wasmInstance);
         })
         .catch(error => {
@@ -581,7 +528,7 @@ class WasmBoyLib {
         // Simply resolve with the input
         resolve(game);
         return;
-      } else if (typeof game === "object" && game.size) {
+      } else if (typeof game === 'object' && game.size) {
         // Read the file object
         // https://www.javascripture.com/FileReader#readAsArrayBuffer_Blob
         const fileReader = new FileReader();
@@ -625,23 +572,20 @@ class WasmBoyLib {
   // Useful for wrapping around .zip files, and using JSZip
   _getGameFromArrayBuffer(fileName, gameBuffer) {
     return new Promise((resolve, reject) => {
-      if (fileName.endsWith(".zip")) {
-        const JSZip = require("jszip");
+      if (fileName.endsWith('.zip')) {
+        const JSZip = require('jszip');
         JSZip.loadAsync(gameBuffer).then(
           zip => {
             // Zip is not an array, but it's proto implements a custom forEach()
             let foundGame = false;
             zip.forEach((relativePath, zipEntry) => {
               if (!foundGame) {
-                if (
-                  relativePath.endsWith(".gb") ||
-                  relativePath.endsWith(".gbc")
-                ) {
+                if (relativePath.endsWith('.gb') || relativePath.endsWith('.gbc')) {
                   // Another function implemented on the proto
                   foundGame = true;
                   zip
                     .file(relativePath)
-                    .async("uint8array")
+                    .async('uint8array')
                     .then(gameBuffer => {
                       resolve(gameBuffer);
                     });
@@ -649,9 +593,7 @@ class WasmBoyLib {
               }
             });
             if (!foundGame) {
-              reject(
-                new Error('The ".zip" did not contain a ".gb" or ".gbc" file!')
-              );
+              reject(new Error('The ".zip" did not contain a ".gb" or ".gbc" file!'));
             }
           },
           error => {
@@ -667,8 +609,8 @@ class WasmBoyLib {
 }
 
 export const WasmBoy = new WasmBoyLib();
-export { WasmBoyAudio } from "./audio/audio";
-export { WasmBoyGraphics } from "./graphics/graphics";
-export { WasmBoyController } from "./controller/controller";
+export { WasmBoyAudio } from './audio/audio';
+export { WasmBoyGraphics } from './graphics/graphics';
+export { WasmBoyController } from './controller/controller';
 // TODO: Remove this, and consolidate public api in Wasmboy
-export { WasmBoyMemory } from "./memory/memory";
+export { WasmBoyMemory } from './memory/memory';
